[refrac] Change how pre-defined problems
- New way to use pre-defined problems
