--- bert_multitask_learning/data_preprocessing/__init__.py
@@ -1,8 +1,0 @@
-from .ner_data import *
-from .cws_data import *
-from .ctb_data import *
-from .pos_data import *
-from .test_data import *
-from .ontonote_data import *
-from .emotion_data import *
-from .preproc_decorator import *
